% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHGCalc4_1_sim_OSR_REDII.R
\name{GHGCalculator4_1_REDII}
\alias{GHGCalculator4_1_REDII}
\title{Title function for calculation of GHG-Emissions and savings of Biodiesel according to BioGrace II}
\usage{
GHGCalculator4_1_REDII(
  Ertrag = 3503,
  Treatment,
  Feuchte = 9,
  Oil_9 = 42.6,
  Nmineral = 142,
  SimNLeach = 0,
  SimN2ON = 0,
  SimResidue = 0,
  sim = FALSE,
  Norgan = 0,
  P2O5 = NULL,
  K2O = NULL,
  CaCO3 = 312.97984872041,
  PSM = 6.609977798,
  Saatgut = 28,
  Diesel = 83.3,
  Field.acidification = NULL,
  ep = 11.7,
  etd = 1.8,
  ResidueCalc = "BioGrace",
  N2OCalcMethod = "Tier1",
  IncludeOrganic = TRUE,
  IncludeResidues = TRUE,
  IncludeSynthetic = TRUE,
  N2OSimMethod = "IPCC",
  IncludeDrying = FALSE,
  IncludeVolatil = TRUE,
  IncludeLeaching = TRUE,
  OtherEm_ha = 0,
  OtherEm_MJ = 0,
  Dis.EF1.F_SN = 0.016,
  Dis.EF1.F_ON = 0.006,
  Dis.EF1.F_CR = 0.006,
  Dis.IPCC.EF4 = 0.014,
  Dis.CAS.Frac_GASF = 0.05,
  Frac_NLeach = 0.24,
  EF.SN.Prod = 3.469,
  Show = list("complete", "eec", "eec.ha", "eec.MJ", "N2O", "N2O.detail.ha",
    "N2O.detail.MJ", "N2O.detail.ha.more", "N2O.detail.MJ.more", "compare", "internal")
)
}
\arguments{
\item{Ertrag}{seed yield in (kg/ha) give as fresh dry matter}

\item{Treatment}{Optional Vector with Treatment/Variant-Codes}

\item{Feuchte}{water content of OSR seed given as (\%), default value is 9\%}

\item{Oil_9}{Oil content of OSR seed given as (\%), at 9\% moisture, default is 42.6}

\item{Nmineral}{mineral N-fertilisation in (kg N/ha), default value is 142 kg N/ha}

\item{SimNLeach}{Optional direct quantification of N-Leaching in (kg N/ha), by using own estimates or from simulation modelling}

\item{SimN2ON}{Optional direct quantification of direct N2O-N Emissions in (kg N/ha), by using own estimates or from simulation modelling}

\item{SimResidue}{Optional direct quantification of N-amount in Crop residues in (kg N/ha), by using own estimates or from simulation modelling}

\item{sim}{logical (vector) for using own estimates/simulated input values}

\item{Norgan}{organic N fertilization in (kg/ha)}

\item{P2O5}{Optional fertilisation rate of phosphorous in (kg P2O5/ha), if not given, fertilisation is assumed to replace P offtake by seeds}

\item{K2O}{Optional fertilisation rate of potassium in (kg K2O/ha), if not given fertilisation is assumed to replace K offtake by seeds}

\item{CaCO3}{Optional rate of liming in (kg CaCO3/ha)}

\item{PSM}{Input of plant protection products in (kg/ha)}

\item{Saatgut}{seed input (kg/ha)}

\item{Diesel}{Diesel use for field operations default value is 83.3 l/ha}

\item{Field.acidification}{needed??}

\item{ep}{emissions from production of biodiesel (g CO2eq/MJ), default is 11.7 g CO2eq/MJ, (typical value - REDII)}

\item{etd}{emissions for transport in the production chain in (g CO2eq/MJ)}

\item{ResidueCalc}{Option for calculation of N from crop residues choices: "Chalon", "IPCC", "DueV", "BioGrace", "SimValue"}

\item{N2OCalcMethod}{Tier 1 uses aggregated emission factors for disaggretated factors use option "custom"}

\item{IncludeOrganic}{inclusion of organic fertiliser N in emission calculation (T/F, yes, no)}

\item{IncludeResidues}{inclusion of crop residues N in emission calculation  (T/F, yes, no)}

\item{IncludeSynthetic}{inclusion of synthetic N in emission calculation  (T/F, yes, no)}

\item{N2OSimMethod}{Options for calculation of direct N2O-emissions "IPCC", # "HUME ,GNOC, DNDC und MODE" SimN2ON nicht 0}

\item{IncludeDrying}{include emissions due to seed drying}

\item{IncludeVolatil}{option (T/F) for inclusion (exclusion) of gaseous emission, default is inclusion T}

\item{IncludeLeaching}{option (T/F) for inclusion (exclusion) of leaching based emission, default is inclusion T}

\item{OtherEm_ha}{other emission in kg CO2eq/ha}

\item{OtherEm_MJ}{other emissions in MJ}

\item{Dis.EF1.F_SN}{Synthetic fertilizer inputs in wet climates IPCC(2019) 0.019 is default value}

\item{Dis.EF1.F_ON}{Other N inputs in wet climates IPCC (2019)  0.006 is default}

\item{Dis.EF1.F_CR}{disaggretated emission factor for N input from crop residues, 0.006 is default}

\item{Dis.IPCC.EF4}{emission factor for N2O emissions from atmospheric deposition of N on soils and water surfaces}

\item{Dis.CAS.Frac_GASF}{disaggregated ammonia emission factor (Volatilization from synthetic fertilizer), 0.05 is default (ammonia-nitrate)}

\item{Frac_NLeach}{Leaching fraction of N-input, default value is 0.24 according to IPCC 2019}

\item{EF.SN.Prod}{Emission for production of synthetic N fertilizer in kg CO2eq/kg N, default is 3.469 for calcium-ammonium nitrate from additinal standard values of BioGraceII}

\item{Show}{Options for output}
}
\value{
data frame with results of calculation, different details can be choosen from
}
\description{
Title function for calculation of GHG-Emissions and savings of Biodiesel according to BioGrace II
}
