% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHGCalc4_1_sim_Weizen_aktuell.R
\name{GHGCalculator_Wheat_4_1}
\alias{GHGCalculator_Wheat_4_1}
\title{Title GHGCalculator_Wheat_4_1}
\usage{
GHGCalculator_Wheat_4_1(
  Ertrag,
  Treatment,
  pH = 6.5,
  Feuchte = 13.5,
  SimNLeach = 0,
  SimN2ON = 0,
  SimResidue = 0,
  sim = FALSE,
  Nmineral = 114,
  Norgan = 0,
  P2O5 = NULL,
  K2O = NULL,
  CaCO3 = NULL,
  Field.acidification = NULL,
  Dataset = NULL,
  PSM = 5.853,
  Saatgut = 36,
  Diesel = 91.3,
  Irrigation_mm = 0,
  Irri.Energy = 0,
  Irri.ef.Energy = 0,
  ep = 15.1,
  etd = 2.2,
  ResidueCalc = "BioGrace",
  N2OCalcMethod = "Tier1",
  N2OSimMethod = "IPCC",
  IncludeOrganic = TRUE,
  IncludeResidues = TRUE,
  IncludeSynthetic = TRUE,
  IncludeDrying = FALSE,
  IncludeVolatil = TRUE,
  IncludeLeaching = TRUE,
  OtherEm_ha = 0,
  OtherEm_MJ = 0,
  Dis.EF1.F_SN = 0.016,
  Dis.EF1.F_ON = 0.006,
  Dis.EF1.F_CR = 0.006,
  Dis.IPCC.EF4 = 0.014,
  Dis.CAS.Frac_GASF = 0.05,
  Frac_NLeach = 0.24,
  EF.SN.Prod = 3.469,
  Show = list("complete", "eec", "eec.ha", "eec.MJ", "N2O", "N2O.detail.ha",
    "N2O.detail.MJ", "N2O.detail.ha.more", "N2O.detail.MJ.more", "compare", "internal")
)
}
\arguments{
\item{Ertrag}{seed yield in (kg/ha) give as fresh dry matter}

\item{Treatment}{Optional Vector with Treatment/Variant-Codes}

\item{Feuchte}{water content of OSR seed given as (\%), default value is 9\%}

\item{SimNLeach}{simulated nitrogen leaching in (kg/ha)}

\item{SimN2ON}{simulated N2O-N emissions in (kg/ha)}

\item{SimResidue}{#'  Optional direct quantification of N-amount in Crop residues in (kg N/ha), by using own estimates or from simulation modelling}

\item{sim}{logical (vector) for using own estimates/simulated input values}

\item{Nmineral}{mineral N-fertilisation in (kg N/ha), default value is 114 kg N/ha}

\item{Norgan}{organic N fertilization in (kg/ha)}

\item{P2O5}{Optional fertilisation rate of phosphorous in (kg P2O5/ha), if not given, fertilisation is assumed to replace P offtake by seeds}

\item{K2O}{Optional fertilisation rate of potassium in (kg K2O/ha), if not given fertilisation is assumed to replace K offtake by seeds}

\item{CaCO3}{Optional rate of liming in (kg CaCO3/ha)}

\item{PSM}{Input of plant protection products in (kg/ha)}

\item{Saatgut}{seed input (kg/ha)}

\item{Diesel}{Diesel use for field operations default value is 83.3 l/ha}

\item{Irrigation_mm}{amount of irrigation in (mm), default value is 0 mm}

\item{Irri.Energy}{energy use for irrigation in (MJ/mm), default value is 0 MJ/mm}

\item{Irri.ef.Energy}{energy use for irrigation in (gCO2eq/MJ used energy), default value is 0 gCO2eq/MJ}

\item{ep}{emissions from production of bioethanol (g CO2eq/MJ), default is 15.1 g CO2eq/MJ,
Red II(2018)other cereals excluding maize ethanol (natural gas  as  process fuel in conventional boiler)}

\item{etd}{emissions from transport and distribution of bioethanol (g CO2eq/MJ), default is 2.2 g CO2eq/MJ,}

\item{ResidueCalc}{Optional method for calculating the amount of N in crop residues, default value is "BioGrace"}

\item{N2OCalcMethod}{"Tier1", "custom" Tier1: EF1=0.01, custom = Verwendung spezifischer Dis.EF1 wie unten gegeben}

\item{N2OSimMethod}{"HUME" ,"GNOC", "DNDC" und "MODE" angeben, wenn SimN2ON nicht 0, steuert die Aufteilung direkter Emissionen auf F_SN, F_ON und F_ER}

\item{Show}{}
}
\value{
data frame with outputs according to the Show parameter
}
\description{
Title GHGCalculator_Wheat_4_1
}
